% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mmpd.R
\name{compute_mmpd}
\alias{compute_mmpd}
\title{Title}
\usage{
compute_mmpd(
  .data,
  gran_x = NULL,
  gran_facet = NULL,
  response = NULL,
  quantile_prob = seq(0.01, 0.99, 0.01),
  dist_ordered = TRUE
)
}
\arguments{
\item{response}{}
}
\value{

}
\description{
Title
}
\examples{
library(tidyverse)
library(gravitas)
library(parallel)
sm <- smart_meter10 \%>\%
filter(customer_id \%in\% c("10017936"))
gran_x = "week_month"
gran_facet = "week_fortnight"
v = compute_mmpd(sm, gran_x, gran_facet,
response = general_supply_kwh)
#month of the year not working in this setup
}
