% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mmpd_panel.R
\name{compute_mmpd_panel}
\alias{compute_mmpd_panel}
\title{compute mmpd for a panel}
\usage{
compute_mmpd_panel(
  .data,
  quantile_prob = seq(0.01, 0.99, 0.01),
  dist_ordered = TRUE,
  nperm = 20
)
}
\arguments{
\item{nperm}{}
}
\value{

}
\description{
compute mmpd for a panel
}
\author{
Sayani07
}
