% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mmpd_panel_grid.R
\name{compute_mmpd_panel_grid}
\alias{compute_mmpd_panel_grid}
\title{}
\usage{
compute_mmpd_panel_grid(
  sim_null_split,
  quantile_prob = seq(0.01, 0.99, 0.01),
  dist_ordered = TRUE,
  nperm = 20
)
}
\value{

}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Sayani Gupta
}
