% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_quantiles.R
\name{compute_quantiles}
\alias{compute_quantiles}
\title{compute quantiles of data across categories}
\usage{
compute_quantiles(sim_panel_data, quantile_prob = seq(0.01, 0.99, 0.01))
}
\arguments{
\item{quantile_prob}{}
}
\value{

}
\description{
compute quantiles of data across categories
}
\author{
Sayani07
}
