% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_wpd_norm.R
\name{compute_wpd_norm}
\alias{compute_wpd_norm}
\title{Title Compute raw pairwise distances}
\usage{
compute_wpd_norm(
  .data,
  gran_x = NULL,
  gran_facet = NULL,
  response = NULL,
  quantile_prob = seq(0.01, 0.99, 0.01),
  dist_ordered = TRUE,
  lambda = 0.67
)
}
\arguments{
\item{.data}{data for which mmpd needs to be calculated}

\item{gran_x}{granularities mapped across x levels}

\item{gran_facet}{granularities mapped across facetss}

\item{response}{univarite response variable}

\item{quantile_prob}{probabilities}

\item{dist_ordered}{if categories are ordered}
}
\value{

}
\description{
Title Compute raw pairwise distances
}
\examples{
library(tidyverse)
library(gravitas)
library(parallel)
sm <- smart_meter10 \%>\%
  filter(customer_id \%in\% c("10017936"))
gran_x <- "week_month"
gran_facet <- "week_fortnight"
v <- compute_wpd_norm(sm, gran_x, gran_facet,
  response = general_supply_kwh, lambda = 0.8
)
# month of the year not working in this setup
}
