% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_all_pairwise.R
\name{distance_all_pairwise}
\alias{distance_all_pairwise}
\title{computing all within and between facet distances between quantile categories given quantile data}
\usage{
distance_all_pairwise(
  sim_panel_quantiles,
  quantile_prob = seq(0.01, 0.99, 0.01),
  dist_ordered = TRUE,
  lambda = 0.67
)
}
\arguments{
\item{sim_panel_quantiles}{quantile data}

\item{quantile_prob}{numeric vector of probabilities with value #'in [0,1]  whose sample quantiles are wanted. Default is set to #' "decile" plot}

\item{dist_ordered}{if categories are ordered}
}
\value{

}
\description{
computing all within and between facet distances between quantile categories given quantile data
}
\author{
Sayani07
}
