% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_panel.R
\name{distance_panel}
\alias{distance_panel}
\title{}
\usage{
distance_panel(
  sim_panel_quantiles,
  quantile_prob = seq(0.01, 0.99, 0.01),
  dist_ordered = TRUE
)
}
\value{

}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Sayani07
}
