% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmpd.R
\name{mmpd}
\alias{mmpd}
\title{}
\usage{
mmpd(sim_panel_data, normx_data, nperm = 20)
}
\value{

}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Sayani07
}
