% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpd.R
\name{mpd}
\alias{mpd}
\title{aggregation across x levels}
\usage{
mpd(sim_panel_data, distance_panel_data, nperm = 20)
}
\arguments{
\item{distance_panel_data}{distance between x levels}

\item{nperm}{number of permutations}
}
\value{

}
\description{
aggregation across x levels
}
\author{
Sayani07
}
