% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpd2.R
\name{mpd2}
\alias{mpd2}
\title{}
\usage{
mpd2(sim_panel_data, distance_panel_data, nperm = 100)
}
\value{

}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Sayani07
}
