% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_mmpd.R
\name{rank_harmony}
\alias{rank_harmony}
\title{Title rank mmpd}
\usage{
rank_harmony(.data, harmony_tbl, response, nperm = 20)
}
\arguments{
\item{.data}{data}

\item{harmony_tbl}{harmony table in the context}

\item{responses}{response variable}
}
\value{

}
\description{
Title rank mmpd
}
