% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_x_for_each_facet.R
\name{shuffle_x_for_each_facet}
\alias{shuffle_x_for_each_facet}
\title{}
\usage{
shuffle_x_for_each_facet(sim_panel_data)
}
\value{

}
\description{
.. content for \description{} no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Sayani Gupta
}
