% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mmpd.R
\name{compute_mmpd}
\alias{compute_mmpd}
\title{Title Compute raw mmpd}
\usage{
compute_mmpd(
  .data,
  gran_x = NULL,
  gran_facet = NULL,
  response = NULL,
  quantile_prob = seq(0.01, 0.99, 0.01),
  dist_ordered = TRUE
)
}
\arguments{
\item{.data}{data for which mmpd needs to be calculated}

\item{gran_x}{granularities mapped across x levels}

\item{gran_facet}{granularities mapped across facetss}

\item{response}{univarite response variable}

\item{quantile_prob}{probabilities}

\item{dist_ordered}{if categories are ordered}
}
\value{

}
\description{
Title Compute raw mmpd
}
\examples{
library(tidyverse)
library(gravitas)
library(parallel)
sm <- smart_meter10 \%>\%
filter(customer_id \%in\% c("10017936"))
gran_x = "week_month"
gran_facet = "week_fortnight"
v = compute_mmpd(sm, gran_x, gran_facet,
response = general_supply_kwh)
#month of the year not working in this setup
}
