% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_panel.R
\name{distance_panel}
\alias{distance_panel}
\title{computing distances between quantile categories}
\usage{
distance_panel(
  sim_panel_quantiles,
  quantile_prob = seq(0.01, 0.99, 0.01),
  dist_ordered = TRUE
)
}
\arguments{
\item{sim_panel_quantiles}{quantile data}

\item{quantile_prob}{numeric vector of probabilities with value #'in [0,1]  whose sample quantiles are wanted. Default is set to #' "decile" plot}

\item{dist_ordered}{if categories are ordered}
}
\value{

}
\description{
computing distances between quantile categories
}
\author{
Sayani07
}
