% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmpd.R
\name{mmpd}
\alias{mmpd}
\title{aggregation across facets}
\usage{
mmpd(sim_panel_data, normx_data, nperm = 20)
}
\arguments{
\item{normx_data}{normalised across x data}

\item{nperm}{number of permutations for estimating paramters
across facets}
}
\value{
mmpd value
}
\description{
aggregation across facets
}
\author{
Sayani07
}
