% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpd.R
\name{mpd}
\alias{mpd}
\title{}
\usage{
mpd(sim_panel_data, distance_panel_data, nperm = 20)
}
\value{

}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Sayani07
}
